% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_spt_att.R
\name{add_spt_att}
\alias{add_spt_att}
\title{Add spatial attribute/s}
\usage{
add_spt_att(
  df,
  lyr,
  att_cols,
  renames = NULL,
  df_x = "long",
  df_y = "lat",
  df_crs = 4326,
  use_crs = NULL,
  maxdist_km = NULL
)
}
\arguments{
\item{df}{Data frame containing coordinate columns.}

\item{lyr}{Character. File path/s for the layer/s with polygons containing the attributes to be added to \code{df}.
Currently limited to accepting geoparquet files read with the sfarrow package.}

\item{att_cols}{Character vector of attribute columns from \code{lyr} to add to \code{df}.}

\item{renames}{Named character vector of attribute columns to rename for the output,
e.g. c("spatial_ind" = "ind", "spatial_ind" = "isIndigenous").}

\item{df_x, df_y}{Character. Name of the columns in \code{df} that have the x and y coordinates.}

\item{df_crs}{Integer. Coordinate reference system that the coordinates in \code{df} correspond to (epsg code).}

\item{use_crs}{Integer. Coordinate reference system (epsg code) to use as the standard for all objects in spatial joins.
Using a projected coordinate system removes the sf warning around assuming coordinates are planar if they're not, and
can sometimes help with spherical geometry issues. If NULL, the crs of the distribution will be used as the standard
(i.e. the \code{df} coords and \code{lyr} will be converted to this crs before joining).}

\item{maxdist_km}{Integer. Maximum distance in kilometres to attribute occurrence records outside \code{lyr}.
Occurrence records that don't fall within the polygons in \code{lyr} but within \code{maxdist_km} will be attributed by
the attributes in the nearest polygon in \code{lyr}.}
}
\value{
Data frame equivalent to \code{df} with specified attributes from \code{lyr}.
}
\description{
Adds attributes from a spatial layer to a data frame,
e.g. add indigenous or vagrant status from a distribution layer to a data frame of taxa occurrences.
}
