% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_environ.R
\name{calc_environ}
\alias{calc_environ}
\title{Calculate a species environment category based on occurrences.}
\usage{
calc_environ(
  presence,
  land_poly,
  thresh = 95,
  out_file,
  force_new = FALSE,
  pres_x = "long",
  pres_y = "lat",
  in_crs = 4326,
  out_crs = in_crs,
  buf = 0
)
}
\arguments{
\item{presence}{Cleaned and filtered dataframe of presences}

\item{land_poly}{sf polygon object indicating the land in the area of interest.}

\item{thresh}{Integer. Threshold for percent records to determine if a species is terrestrial, marine or both.}

\item{out_file}{Character. File path for the results files to be saved.}

\item{force_new}{Logical. If \code{out_file} exists, recreate it?}

\item{pres_x, pres_y}{Character. Name of the columns in \code{presence} that have
the x and y coordinates.}

\item{in_crs}{epsg code for coordinates in \code{presence}.}

\item{out_crs}{epsg code for coordinates in output mcp.}

\item{buf}{Distance in metres to buffer the \code{land_poly}.}
}
\value{
data frame saved as rds to \code{out_file} path containing environ and land_pc (percent land overlap ) fields for the taxa.
}
\description{
!Experimental! In future this can also be based on distributions.
}
