% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_grd.R
\name{make_grd}
\alias{make_grd}
\title{Make grid around occurrence records}
\usage{
make_grd(
  presence,
  out_file,
  force_new = FALSE,
  pres_x = "long",
  pres_y = "lat",
  pres_taxa = "taxa",
  in_crs = 4326,
  out_crs = in_crs,
  use_ConR = FALSE,
  cell_size = 2,
  num_rast_pos = 0
)
}
\arguments{
\item{presence}{Cleaned and filtered dataframe of presences.}

\item{out_file}{Character. Path for the mcp to be saved. Will be saved by
\code{sfarrow::st_write_parquet()}. Currently will not work very well with any
full stop in the path. Other file types are changed to .parquet.}

\item{force_new}{Logical. If \code{out_file} exists, recreate it?}

\item{pres_x, pres_y}{Character. Name of the columns in \code{presence} that have
the x and y coordinates.}

\item{pres_taxa}{Character. Name of the column in \code{presence} that
contains the taxa names. Needed for ConR (see below).}

\item{in_crs}{epsg code for coordinates in \code{presence}.}

\item{out_crs}{epsg code for coordinates in output grid.}

\item{use_ConR}{Use the ConR package for creating the grid.
This uses ConR::AOO.computing to find the grid with the lowest number of cells,
but can be very slow for common taxa at national level.
Otherwise, sf::st_make_grid is used. Default is FALSE.}

\item{cell_size}{Grid cell size in km,
e.g. the standard IUCN redlist cell size is 2 km (i.e. a 2 x 2 km square),
which is the default.}

\item{num_rast_pos}{Number of raster start positions when ConR = TRUE.
See nbe.rep.rast.AOO parameter in ConR::AOO.computing for details.}
}
\value{
sf. \code{out_file} saved.
}
\description{
Usually used as Area of Occupancy (AOO) estimate for a taxa.
}
