% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reg_cont.R
\name{reg_cont}
\alias{reg_cont}
\title{Regional contribution to national range (AOO and EOO)}
\usage{
reg_cont(
  taxa,
  presence,
  out_dir,
  mcp_file = NULL,
  grd_file = NULL,
  region_bound,
  force_new = FALSE,
  pres_x = "long",
  pres_y = "lat",
  in_crs = 4326,
  out_crs = in_crs,
  nearest_pop = FALSE,
  use_ConR = TRUE,
  cell_size = 2,
  num_rast_pos = 0,
  clip = use_clip
)
}
\arguments{
\item{taxa}{Character. Taxa name for use in ConR and outputs.}

\item{presence}{Cleaned and filtered dataframe of presences.}

\item{out_dir}{Character. Directory path for the results files to be saved (summary_stats.rds and mcp_noreg.parquet).
mcp_noreg will be saved by \code{sfarrow::st_write_parquet()}. Currently will not work very well with any
full stop in the path. Other file types are changed to .parquet.}

\item{mcp_file}{Character. Path for the national mcp. If the file exists it will be read in,
otherwise it will be created using envSDM::make_mcp.}

\item{grd_file}{Character. Path for the national grid. If the file exists it will be read in,
otherwise it will be created using make_grd from envRange.}

\item{region_bound}{sf object. Region boundary polygon.}

\item{force_new}{Logical. If any of the results files in the \code{out_dir} exists, recreate it?}

\item{pres_x, pres_y}{Character. Name of the columns in \code{presence} that have
the x and y coordinates.}

\item{in_crs}{epsg code for coordinates in \code{presence}.}

\item{out_crs}{epsg code for coordinates in output grid.}

\item{nearest_pop}{Logical. Calculate distance to nearest extra-regional population?}

\item{use_ConR, cell_size, num_rast_pos}{Passed to make_grd if national grid doesn't exist or force_new = TRUE.}

\item{clip}{Passed to make_mcp if national mcp doesn't exist or force_new = TRUE.}
}
\value{
summary_stats.rds with aoo and eoo stats,
and mcp_noreg.parquet with the mcp excluding regional presences.
}
\description{
Regional contribution to national range (AOO and EOO)
}
