% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusts.R
\name{clusts}
\alias{clusts}
\title{Identify clusters of occurrence points based on distance.}
\usage{
clusts(presence, h, ...)
}
\arguments{
\item{presence}{Cleaned and filtered dataframe of presences}

\item{h}{Distance used to cluster points, i.e. points within individual clusters
will not be further apart than this distance.}
}
\value{
data frame saved as rds to \code{out_file} path containing environ and land_pc (percent land overlap ) fields for the taxa.
}
\description{
Can be used as a precursor to make_mcp or make_ahull for creating refined distribution polygons
for species with geographically separated populations/subspecies.
}
