% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reg_cont.R
\name{reg_cont}
\alias{reg_cont}
\title{Regional contribution to range (AOO and EOO)}
\usage{
reg_cont(
  taxa,
  presence,
  out_dir,
  mcp_file = NULL,
  grd_file = NULL,
  region_bound,
  force_new = FALSE,
  pres_x = "long",
  pres_y = "lat",
  in_crs = 4326,
  out_crs = in_crs,
  nearest_pop = FALSE,
  clip = NULL
)
}
\arguments{
\item{taxa}{Character. Taxa name for use in ConR and outputs.}

\item{presence}{Cleaned and filtered dataframe of presences.}

\item{out_dir}{Character. Directory path for the results files to be saved (summary_stats.rds and mcp_noreg.parquet).
mcp_noreg will be saved by \code{sfarrow::st_write_parquet()}. Currently will not work very well with any
full stop in the path. Other file types are changed to .parquet.}

\item{mcp_file}{Character. Path for the range mcp.}

\item{grd_file}{Character. Path for the range grid.}

\item{region_bound}{sf object. Region boundary polygon.}

\item{force_new}{Logical. If any of the results files in the \code{out_dir} exists, recreate it?}

\item{pres_x, pres_y}{Character. Name of the columns in \code{presence} that have
the x and y coordinates.}

\item{in_crs}{epsg code for coordinates in \code{presence}.}

\item{out_crs}{epsg code for coordinates in output 'no region' mcp.}

\item{nearest_pop}{Logical. Calculate distance to nearest extra-regional population?}

\item{clip}{sf to clip the 'no region' mcp for comparing the impact of removing regional presences.}
}
\value{
summary_stats.rds with aoo and eoo stats,
and mcp_noreg.parquet with the mcp excluding regional presences.
}
\description{
Typically used to determine the contribution of a region to the national range of a species,
where a 'region' is a smaller area within the range.
}
