% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reg_taxa.R
\name{reg_taxa}
\alias{reg_taxa}
\title{Return relevant taxa to a region of interest based on occurrences and or distributions.}
\usage{
reg_taxa(
  presence,
  distrib_file = NA,
  region_bound,
  taxa,
  pres_x = "long",
  pres_y = "lat",
  pres_crs = 4326,
  use_crs = NULL,
  buf = 0,
  taxonomy,
  remove = FALSE
)
}
\arguments{
\item{presence}{Cleaned and filtered dataframe of presences.}

\item{distrib_file}{Character. File path/s for the relevant distributions per taxa. If no relevant distribution, then use NA.
Currently, only parquet files are accepted.
Use sfarrow::st_write_parquet to write sf objects to parquet.}

\item{region_bound}{sf object. Region boundary polygon.}

\item{taxa}{Character. Name of the taxa. Needed for joining to taxonomy to obtain species for trinomials (see below).}

\item{pres_x, pres_y}{Character. Name of the columns in \code{presence} that have
the x and y coordinates.}

\item{use_crs}{Integer. Coordinate reference system (epsg code) to use as the standard for all objects in spatial joins or intersections.
Using a projected coordinate system removes the sf warning around assuming coordinates are planar if they're not, and
can sometimes help with spherical geometry issues. If NULL, the crs of the region_bound will be used as the standard
(i.e. the presence coords will be converted to this crs before joining or intersecting).}

\item{buf}{Distance in metres to buffer the \code{region_bound}.}

\item{taxonomy}{Taxonomy object returned by envClean::make_taxonomy in relation to the taxa.
Must have subspecies level taxonomy.}

\item{remove}{Logical. Remove out of region taxa?}

\item{in_crs}{epsg code for coordinates in \code{presence}.}
}
\value{
Data frame with species column, subspecies column, and 'pres_dist' & 'distrib_dist' columns indicating
distance in metres from region to closest presence and distribution respectively. If remove == TRUE, rows with
out of region subspecies will not appear.
}
\description{
Useful for species level analyses seeking to link to relevant trinomials (subspecies, races, varieties, forms, etc),
where there may be other irrelevant trinomials occurring outside the region of interest.
}
