% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reg_taxa.R
\name{reg_taxa}
\alias{reg_taxa}
\title{Return relevant taxa to a region of interest based on occurrences and or distributions.}
\usage{
reg_taxa(
  presence,
  out_dir,
  region_bound,
  force_new = FALSE,
  pres_x = "long",
  pres_y = "lat",
  in_crs = 4326,
  out_crs = in_crs,
  buf = 0
)
}
\arguments{
\item{presence}{Cleaned and filtered dataframe of presences.}

\item{out_dir}{Character. Directory path for the results files to be saved (summary_stats.rds and mcp_noreg.parquet).
mcp_noreg will be saved by \code{sfarrow::st_write_parquet()}. Currently will not work very well with any
full stop in the path. Other file types are changed to .parquet.}

\item{region_bound}{sf object. Region boundary polygon.}

\item{force_new}{Logical. If any of the results files in the \code{out_dir} exists, recreate it?}

\item{pres_x, pres_y}{Character. Name of the columns in \code{presence} that have
the x and y coordinates.}

\item{in_crs}{epsg code for coordinates in \code{presence}.}

\item{out_crs}{epsg code for coordinates in output grid.}

\item{buf}{Distance in metres to buffer the \code{region_bound}.}
}
\value{
summary_stats.rds with aoo and eoo stats,
and mcp_noreg.parquet with the mcp excluding regional presences.
}
\description{
Useful for species level analyses seeking to link to relevant trinomials (subspecies, races, varieties, forms, etc),
where there may be other irrelevant trinomials occurring outside the region of interest.
}
